Check for import recursion using MAX_STACK_SIZE
The current check for infinite recursive imports is currently too simple
and gets in the way of practical use cases for importing macros in both
base templates and partials, for example.

This replaces it with an "imperfect" stack size check, with a value we
felt, if exceeded, probably means something is wrong.